testing client

+++ client/backend/models/group.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Boolean, String
+from sqlalchemy import Column, Integer, Boolean, String, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 Base = declarative_base()
 from sqlalchemy.orm import relationship
@@ -9,7 +9,7 @@ class Group(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, index=True)
     hierarchy = Column(Boolean)
-    creator = Column(Integer, nullable=False)
+    creator = Column(BigInteger, nullable=False)
     users = relationship("User", secondary=association_table, back_populates="groups")
