+++ client/backend/models/notification.py
@@ -1,5 +1,5 @@
 from sqlalchemy import UniqueConstraint
-from sqlalchemy import Column, Integer, String, ForeignKey, DATETIME
+from sqlalchemy import Column, Integer, String, ForeignKey, DATETIME, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 Base = declarative_base()
 
@@ -7,7 +7,7 @@ class Notification(Base):
     __tablename__ = "notifications"
     
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey('users.id'), nullable=False)
+    user_id = Column(BigInteger, ForeignKey('users.id'), nullable=False)
     text = Column(String, nullable=False)
     time = Column(DATETIME, nullable=False)
